requirements.txt
streamlit>=1.32.2
python-docx>=1.1.0
pillow>=11.0.0
groq>=0.9.0
pytesseract
pdf2image
fpdf2
mammoth
opencv-python 
numpy